Add an SSE 4.2 hotpath using the crc extensions.
SSE 4.2 has an awesome set of x86 extensions for CRC functions. This
commit makes use of them if hashmap.h was built with a compiler that
supports SSE 4.2 or above. To ensure that users could mix code that
used the SSE 4.2 crc functions and the scalar fallback code in multiple
compilation units of their executable, the fallback CRC function was
changed to mimic the polynomial that Intel uses for the CRC functions.

I also fixed a bug with the `initial_size` of `hashmap_create` it is
expected to be a power of 2 value, and that wasn't recorded previously.
Now the create function will return an error if the size isn't a power
of two.